Allow tagging an ENI with "no detach"
And make it so ipamd won't detach any such ENI.

This also regenerates the mocks with a slightly newer mockgen.

The use-case for the behavior is for a user who wishes to attach an ENI
to an ec2 instance that they manage, but which the cni plugin will still
use to aquire additional IPs as needed.

Specifically, this allows an ENI's primary IP to be used for normal
traffic without fear of the ENI later being detached by this CNI code.