Add a rule to let the primary ip on secondary ENIs egress
Without this rule, the first rule that matches traffic coming out of
eth1 is 'from all lookup main', which has a default route going out
eth0.

This means that traffic trying to egress from eth1's primary ip will go
out the wrong interface, and the primary ip can't reasonably be used
(e.g. to listen on or such).

This adds a rule for that ip specifically which should fix this. It
intentionally uses a priority such that the fwmark rule will happen
first, so HostPort traffic should still end up on eth0 as desired.